Implement a workable solution for issue #345
Also tackle legacy device support, downgrade minSdkVersion to 18 from 19
unless someone can tell me that there is a reason the code base must be
on API 19.  Note: I've been working at this to be able to continue to
maintain at least one Github client working with BB10 (Blackberry OS10),
which has ForkHub as its sole Github client in the app store
BlackberryWorld (also in Amazon store), however this app stopped working
about a year or so back, and this should provide a fix.  Work with
upstream maintainers to sort out cleanest fix as this only takes into
account a build specific to the legacy devices.